+++ src/main/java/dev/marksman/enhancediterables/ImmutableIterable.java
@@ -408,4 +408,15 @@ public interface ImmutableIterable<A> extends EnhancedIterable<A> {
         }
     }
 
+    /**
+     * Returns an infinite {@code ImmutableNonEmptyIterable} that repeatedly iterates a given element.
+     *
+     * @param element the value to repeat
+     * @param <A>     the element type
+     * @return an {@code ImmutableNonEmptyIterable<A>}
+     */
+    static <A> ImmutableNonEmptyIterable<A> repeat(A element) {
+        return EnhancedIterables.repeat(element);
+    }
+
 }