+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
-No changes yet
+### Added
+- `repeat` constructors
+
+### Changed
+- Singletons now require less memory
 
 ## [1.0.4] - 2019-10-01
 ###