+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `cycle` method for `FiniteIterable`s
 
 ### Changed
+- Upgrade lambda to 5.1.0
 - Singletons now require less memory
 
 ## [1.0.4] - 2019-10-01